Fix loading bug caused by adding the filename
`cluster_2_sector` was being added to AX as if it was a full word,
when the variable itself was a byte. By coincidence, the next data was
all zeroes, and since it is little-endian there wasn't an issue. But
the addition of another variable caused this to break. This makes the
variable explicitly a word in order to facilitate the arithmetic more
easily and avoid this bug.