Basic memory footprint reduction
This moves the storage of the FAT and root directory listing to the
1.5KiB before the boot sector (using up what was previously stack
space). Only two sectors of the FAT are loaded and only one sector of
the RDL is loaded.

For the first 14 entries in a defragmented filesystem, everything
should just work the same as before without any further
modification (so long as the files are relatively small). Enough of
the FAT should be loaded to correctly follow the cluster chain, and
the first 14 entries of the root directory have already been loaded.

The next step from here is to modify the existing routines to change
which 1KiB section of the FAT is loaded according to which cluster is
being loaded, and to cycle through the sectors of the root directory
if the first 14 entries do not contain the file we need.